requirements.txt
undetected-chromedriver~=3.5.5
pandas~=2.2.2
beautifulsoup4~=4.12.3
webdriver-manager~=4.0.2
selenium_stealth~=1.0.6
fake_useragent~=2.0.3